changed the oled and macros
