+++ keyboards/crkbd/keymaps/marleydepew/keymap.c
@@ -1,23 +1,6 @@
-/*
-Copyright 2019 @foostan
-Copyright 2020 Drashna Jaelre <@drashna>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
 #include QMK_KEYBOARD_H
 #include "../../macro.h"
+#include "../../oled.h"
 
 enum layer_names {
     _COLEMAK_DH,
@@ -30,13 +13,17 @@ enum layer_names {
 };
 
 enum custom_keycodes {
-    MACRO_0 = SAFE_RANGE,
-    MACRO_1,
-    MACRO_2,
+    MACRO_V = SAFE_RANGE,
+    MACRO_A,
+    MACRO_C,
     MACRO_TM,
     MACRO_TH,
     MACRO_WM,
-    MACRO_WL
+    MACRO_WL,
+    MACRO_1,
+    MACRO_2,
+    MACRO_3,
+    MACRO_4
 };
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
@@ -102,11 +89,11 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
     [_MOUSE] = LAYOUT_split_3x6_3(                                        // Mouse Layer
     // |-----------+-----------+-----------+-----------+-----------+-----------|   |-----------+-----------+-----------+-----------+-----------+-----------|
-        MACRO_0,    KC_WH_L,    KC_WH_U,    KC_WH_D,    KC_WH_R,    XXXXXXX,        XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,
+        MACRO_V,    KC_WH_L,    KC_WH_U,    KC_WH_D,    KC_WH_R,    XXXXXXX,        XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,
     // |-----------+-----------+-----------+-----------+-----------+-----------|   |-----------+-----------+-----------+-----------+-----------+-----------|
-        MACRO_1,    KC_MS_L,    KC_MS_U,    KC_MS_D,    KC_MS_R,    LSFT(KC_F3),    XXXXXXX,    KC_RCTL,    KC_RSFT,    KC_RALT,    KC_RGUI,    XXXXXXX,
+        MACRO_A,    KC_MS_L,    KC_MS_U,    KC_MS_D,    KC_MS_R,    LSFT(KC_F3),    XXXXXXX,    KC_RCTL,    KC_RSFT,    KC_RALT,    KC_RGUI,    XXXXXXX,
     // |-----------+-----------+-----------+-----------+-----------+-----------|   |-----------+-----------+-----------+-----------+-----------+-----------|
-        MACRO_2,    XXXXXXX,    XXXXXXX,    KC_BTN1,    KC_BTN2,    KC_BTN3,        XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,
+        MACRO_C,    XXXXXXX,    XXXXXXX,    KC_BTN1,    KC_BTN2,    KC_BTN3,        XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,
     // |-----------+-----------+-----------+-----------+-----------+-----------|   |-----------+-----------+-----------+-----------+-----------+-----------|
                                             KC_ACL0,    KC_ACL1,    KC_ACL2,        XXXXXXX,    XXXXXXX,    XXXXXXX
     //                                     |-----------+-----------+-----------|   |-----------+-----------+-----------|
@@ -125,146 +112,6 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   )
 };
 
-// The next section is only applicable if OLEDs are enabled.
-#ifdef OLED_ENABLE
-
-oled_rotation_t oled_init_user(oled_rotation_t rotation) {
-
-    if (is_keyboard_master()) {
-        return OLED_ROTATION_270;
-    } else {
-        return OLED_ROTATION_0;
-    }
-}
-
-void oled_render_logo(void) {
-    // clang-format off
-    static const char PROGMEM marley_katakana[] = {
-        // https://javl.github.io/image2cpp/
-        // Code Output: Plain Bytes; Draw Mode: Vertical; Rotate Image: 0 degrees
-        // 'marley katakana_rotated.jpg'; 128x32px   
-        // (source image was rotated 90 degrees to lay horizontally
-        // with the bottom of the image on the left and the top of the image on the right)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0xc0, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0xf0, 0xf0, 0xf0, 0xf0, 0xe0, 0xc0, 0xc0, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x80, 0xc0, 0xe0, 0xe0, 0xc0, 0xc0, 0xc0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 
-        0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xe0, 0xc0, 0x80, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0xc0, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x0e, 0x0f, 0x0f, 0x1f, 
-        0x1f, 0x1f, 0x3f, 0x3f, 0x0f, 0x0f, 0x0f, 0x0f, 0x0f, 0x1f, 0x1f, 0x0f, 0x0f, 0x03, 0x01, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0xc0, 0xe0, 0xc0, 0xc0, 0xc0, 0xe0, 0xe0, 
-        0xe0, 0xe0, 0xe0, 0xe0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xe0, 0x80, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xc0, 0xe0, 0xf0, 0xf8, 
-        0x38, 0x00, 0x00, 0x01, 0x0f, 0x3f, 0xff, 0xff, 0xf1, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x01, 0x07, 0x0f, 0x07, 0x07, 0x03, 0x03, 0x03, 0x01, 0x01, 0x01, 0x01, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 0x01, 0x01, 0x03, 0x03, 0x03, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x07, 0x0e, 0x1c, 0x3c, 0x78, 0xf0, 0xf0, 0xf0, 0xf0, 0xe0, 
-        0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xf0, 0xf4, 0xfc, 0x78, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x0f, 0x07, 0x07, 0x07, 0x03, 0x03, 0x01, 
-        0x01, 0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 0x01, 0x01, 0x03, 0x03, 0x03, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x70, 0x7c, 0x7e, 0x3f, 0x3f, 0x1f, 0x07, 0x0f, 
-        0x3e, 0x7c, 0xf8, 0xf0, 0xe0, 0xe0, 0xc1, 0xcf, 0xff, 0xff, 0xf8, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 
-        0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
-        0x00, 0x00, 0x00, 0x01, 0x03, 0x07, 0x0f, 0x0f, 0x07, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-    };
-    // clang-format on
-    oled_write_raw_P(marley_katakana, sizeof(marley_katakana));
-}
-
-void oled_render_status(void) {
-
-    // Print layer status
-    oled_set_cursor(0, 0);
-    oled_write("LAYER", false);
-
-    oled_set_cursor(0, 2);
-    switch (get_highest_layer(layer_state)) {
-
-        case _COLEMAK_DH:
-            oled_write("COLMK", false);
-            break;
-
-        case _NUMBER:
-            oled_write("NUMBR", false);
-            break;
-
-        case _SYMBOL:
-            oled_write("SYMBL ", false);
-            break;
-
-        case _FUNCTION:
-            oled_write("FNCTN", false);
-            break;
-
-        case _NAVIGATION:
-            oled_write("NVGTN", false);
-            break;
-
-        case _MOUSE:
-            oled_write("MOUSE", false);
-            break;
-
-        case _RGB:
-            oled_write("RGBLT", false);
-            break;
-
-    default:
-            oled_write("UNDEF", false);
-
-    }
-
-    led_t led_usb_state = host_keyboard_led_state();
-
-    // print num lock status
-    oled_set_cursor(0, 8);
-    if(led_usb_state.num_lock) {
-        oled_write("NUMLK", true);
-    } else {
-        oled_write("     ", false);
-    }
-    
-    // print scroll lock status
-    oled_set_cursor(0, 10);
-    if (led_usb_state.scroll_lock) {
-        oled_write("SCRLK", true);
-    } else {
-        oled_write("     ", false);
-    }
-
-    // print caps lock status
-    oled_set_cursor(0, 12);
-    if (led_usb_state.caps_lock) {
-        oled_write("CPSLK", true);
-    } else {
-        oled_write("     ", false);
-    }
-
-}
-
-bool oled_task_user(void) {
-    
-    if (is_keyboard_master()) {
-        oled_render_status();
-    } else {
-        oled_render_logo();
-    }
-
-    return false;
-}
-
-#endif // End OLED_ENABLE
-
 // These are my custom overrides for the shift keys
 #ifdef KEY_OVERRIDE_ENABLE
 
@@ -279,12 +126,6 @@ const key_override_t nine_override = ko_make_with_layers(MOD_MASK_SHIFT, KC_9, K
 
 // This globally defines all key overrides to be used
 // const key_override_t **key_overrides = (const key_override_t *[]){
-//     &left_parenth_override,
-//     &right_parenth_override,
-//     &nine_override,
-//     NULL // Null terminate the array of overrides
-// };
-
 const key_override_t *key_overrides[] = {
     &left_parenth_override,
     &right_parenth_override,
@@ -300,21 +141,41 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
     // This is for the macros
     switch (keycode) {
         
-        case MACRO_0:
+        case MACRO_V:
             if (record->event.pressed) {
-                // when keycode MACRO_0 is pressed
+                // when keycode MACRO_V is pressed
                 // The string is defined in "../../macro.h"
-                SEND_STRING(M0_STRING);
+                SEND_STRING(MV_STRING);
             } else {
-                // when keycode MACRO_0 is released
+                // when keycode MACRO_V is released
+            }
+            break;
+
+        case MACRO_A:
+            if (record->event.pressed) {
+                // when keycode MACRO_A is pressed
+                // The string is defined in "../../macro.h"
+                SEND_STRING(MA_STRING);
+            } else {
+                // when keycode MACRO_A is released
+            }
+            break;
+
+        case MACRO_C:
+            if (record->event.pressed) {
+                // when keycode MACRO_C is pressed
+                // The string is defined in "../../macro.h"
+                SEND_STRING(MC_STRING);
+            } else {
+                // when keycode MACRO_C is released
             }
             break;
 
         case MACRO_1:
             if (record->event.pressed) {
                 // when keycode MACRO_1 is pressed
-                // The string is defined in "../../macro.h"
-                SEND_STRING(M1_STRING);
+                // Open Python in Command Prompt
+                SEND_STRING(SS_RGUI('r') SS_DELAY(500) 'cmd.exe' SS_DELAY(1000) 'python.exe')
             } else {
                 // when keycode MACRO_1 is released
             }
@@ -324,9 +185,29 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
             if (record->event.pressed) {
                 // when keycode MACRO_2 is pressed
                 // The string is defined in "../../macro.h"
-                SEND_STRING(M2_STRING);
+                SEND_STRING(SS_RGUI('r') SS_DELAY(500) 'chrome.exe' SS_DELAY(1000) 'https://amgensbx-rim-test.veevavault.com')
             } else {
                 // when keycode MACRO_2 is released
+            }
+            break;
+
+        case MACRO_3:
+            if (record->event.pressed) {
+                // when keycode MACRO_3 is pressed
+                // The string is defined in "../../macro.h"
+                SEND_STRING(M3_STRING);
+            } else {
+                // when keycode MACRO_3 is released
+            }
+            break;
+
+        case MACRO_4:
+            if (record->event.pressed) {
+                // when keycode MACRO_4 is pressed
+                // The string is defined in "../../macro.h"
+                SEND_STRING(M4_STRING);
+            } else {
+                // when keycode MACRO_4 is released
             }
             break;
